Añadi politicas de acceso a usuarios, roles y permisos

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -15,17 +15,24 @@ class PermissionController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
-        $permisos=Permission::all();
-        return view('admin.permisos.index',[
-            'permisos'=>$permisos]);
-    }
-    public function store(PermisoRequest $request)
-    {
-        $permiso=Permission::create($request->validated());
-        return back()->with('mensaje','Se ha agregado un nuevo permiso');
-    }
+    public function __construct(){
+      $this->middleware('auth');
+  }
+  public function index()
+  {
+    //Aplicamos Politica de Acceso al metodo correspondiente
+    $this->authorize('view',new Permission);
+    $permisos=Permission::all();
+    return view('admin.permisos.index',[
+        'permisos'=>$permisos]);
+}
+public function store(PermisoRequest $request)
+{
+     //Aplicamos Politica de Acceso al metodo correspondiente
+    $this->authorize('create',new Permission);
+    $permiso=Permission::create($request->validated());
+    return back()->with('mensaje','Se ha agregado un nuevo permiso');
+}
 
     /**
      * Display the specified resource.
@@ -40,6 +47,8 @@ class PermissionController extends Controller
 
     public function update(PermisoRequest $request, Permission $permiso)
     {
+        //Aplicamos Politica de Acceso al metodo correspondiente
+        $this->authorize('update',$permiso);
         $permiso->update($request->validated());
         return back()->with('mensaje','Se ha actualizado el permiso correctamente');
     }
@@ -54,6 +63,8 @@ class PermissionController extends Controller
     {
         $id= Crypt::decryptString($id);
         $permiso=Permission::findOrFail($id);
+         //Aplicamos Politica de Acceso al metodo correspondiente
+        $this->authorize('delete',$permiso);
         $permiso->delete();
         return back()->with('mensaje','Se ha eliminado el permiso');
     }
